1) mucke-backend: initial form of facet search for demo 2) mucke-backend: interactive properties file for interactive mode (UI mode) 3) mucke-backend: Index now built with StringFields, bug fix from Textfields 4) mucke-backend: cleanup, removed old ImageSearcher and TextSearcher class 5) mucke-ui: added new Result class to EditorBean 6) mucke-ui: modified xhtml to display path based on image location from the facet index, and caption as the title 7) mucke-ui: Full query language 8) Added credibility checkbox on search interface (however, not working yet) 9) Build backend pom back to Lucene 4.2.1 since 4.7 causes Lucene Luke not to work any more (but it required for index inspection).

+++ mucke-backend/src/main/java/at/tuwien/mucke/plugin/prototypetest/crediblity/CSVCredibilityReader.java
@@ -36,7 +36,7 @@ public class CSVCredibilityReader implements CredibilityReader {
 
         // add part 1-3 as id, query terms, and narrative
         String userid = parts[0].trim();
-        double score = Double.parseDouble(parts[1].trim());
+        float score = Float.parseFloat(parts[1].trim());
 
         return new User(userid, score);
     }
